DAOS-10281 test: DAOS interoperability test formatted 2.2 storage downgrade
Included tests:
DAOS-10282: DAOS interoperability negative test pool upgrade with system stop
DAOS-10279: DAOS interoperability test pool upgrade interrupt and resume after server upgraded
DAOS-10288: DAOS interoperability test DAOS system and pool upgrade basic with 8 servers

Steps:
    (1)Setup and show rpm, dmg and daos versions on all hosts
    (2)Create pool, container and pool attributes
    (3)Setup and run IOR
        (3.a)DFS
        (3.b)HDF5
        (3.c)POSIX symlink to a file
    (4)Dmg system stop
    (5)Upgrade RPMs to specified new version
    (6)Restart servers
    (7)Restart agent
        verify pool attributes
        verify IOR data integrity after upgraded
        (7.a)DFS
        (7.b)HDF5
        (7.c)POSIX symlink to a file
    (8)Dmg pool get-prop after RPMs upgraded before Pool upgraded
    (9)Dmg pool upgrade and verification after RPMs upgraded
        (9.a)Enable fault injection during pool upgrade
        (9.b)Normal pool upgrade without fault injection
    (10)Create new pool after rpms Upgraded
    (11)Downgrade and cleanup
    (12)Restart servers and agent

Test-tag: interop,-manual
Required-githooks: true
