+++ src/tests/ftest/interoperability/down_grade.yaml
@@ -30,7 +30,7 @@ ior:
   client_processes:
     np: 2
   test_file: daos:/testFile
-  api: DFS   # POSIX HDF5
+  api: DFS   # HDF5, POSIX
   dfs_oclass: "SX"  # "RP_2GX"
   transfer_size: '1M'
   block_size: '128M'
@@ -43,6 +43,5 @@ interop:
   # Example of upgrade/downgrade RPMs from local tar file
   # upgrade_rpms: ["/home/dinghwah/RPM/2.1.104/daos-2.1.104-1.el8.x86_64.rpm"]
   # downgrade_rpms: ["/home/dinghwah/RPM/2.0.3.rc4.signed/daos-2.0.3-5.el8.x86_64.rpm"]
-  # upgrade_repo: "/etc/yum.repos.d/test23-101.repo"
   downgrade_repo: "/etc/yum.repos.d/test22-1.repo"
   upgrade_repo: "/etc/yum.repos.d/test23-101.repo"
